Refactor expression parser modules
