+++ include/parser_expr_literal.h
@@ -1,0 +1,10 @@
+#ifndef VC_PARSER_EXPR_LITERAL_H
+#define VC_PARSER_EXPR_LITERAL_H
+
+#include "parser.h"
+#include "parser_types.h"
+#include "ast_expr.h"
+
+expr_t *parse_literal(parser_t *p);
+
+#endif /* VC_PARSER_EXPR_LITERAL_H */
