+++ include/parser_expr_primary.h
@@ -7,5 +7,7 @@
 
 /* Parse a primary expression including unary operators */
 expr_t *parse_primary(parser_t *p);
+/* Helper used by postfix parser */
+expr_t *parse_base_term(parser_t *p);
 
 #endif /* VC_PARSER_EXPR_PRIMARY_H */
